requirements.txt
pyautogui
pycaw
comtypes
opencv-python
mediapipe
typesense
python-vlc
numpy

